+++ packages/editor/src/plugins/index.ts
@@ -1,3 +1,4 @@
 export * from './selection-change';
 export * from './plugin';
 export * from './undo';
+export * from './selection';