+++ packages/editor/src/utils/ranges.ts
@@ -10,14 +10,6 @@ export function removeRanges() {
   }
 }
 
-export function saveSelection() {
-  const selection = window.getSelection();
-  if (!selection) {
-    return;
-  }
-  return selection.rangeCount === 0 ? undefined : selection.getRangeAt(0);
-}
-
 export function restoreSelection(range?: Range) {
   const selection = window.getSelection();
   if (!selection || !range) {
@@ -65,3 +57,21 @@ export function expandToTag(node: Node) {
   range.selectNodeContents(node);
   selection.addRange(range);
 }
+
+// FIXME: figure how to detect select all
+export const isEditableSelectAll = () => {
+  const sel = getSelection();
+  if (!sel) return;
+  if (sel.type === 'Caret') {
+    if (!sel.anchorOffset && sel.isCollapsed) {
+      return true;
+    }
+    return false;
+  }
+  if (sel.type === 'Range') {
+    if (!sel.anchorOffset) {
+      return true;
+    }
+    return false;
+  }
+};